// apps/client/src/features/project/__generated__/projectTasks_Connection.graphql.ts
export type TaskStatus = "COMPLETED" | "IN_PROGRESS" | "NOT_STARTED" | "%future added value";
import { FragmentRefs } from "relay-runtime";
export type projectTasks_Connection$data = {
  readonly id: string;
  readonly taskEdge: {
    readonly edges: ReadonlyArray<{
      readonly cursor: string;
      readonly node: {
        readonly description: string | null | undefined;
        readonly id: string;
        readonly status: TaskStatus | null | undefined;
        readonly title: string | null | undefined;
      } | null | undefined;
    } | null | undefined> | null | undefined;
    readonly pageInfo: {
      readonly endCursor: string | null | undefined;
      readonly hasNextPage: boolean;
      readonly hasPreviousPage: boolean;
      readonly startCursor: string | null | undefined;
    };
  };
  readonly " $fragmentType": "projectTasks_Connection";
};
export type projectTasks_Connection$key = {
  readonly " $data"?: projectTasks_Connection$data;
  readonly " $fragmentSpreads": FragmentRefs<"projectTasks_Connection">;
};